Move event related classes into their own package
