+++ app/src/androidTest/java/com/github/onedirection/testdatabase/ConcreteDatabaseTest.java
@@ -2,7 +2,7 @@ package com.github.onedirection.testdatabase;
 
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 
-import com.github.onedirection.Event;
+import com.github.onedirection.events.Event;
 import com.github.onedirection.database.ConcreteDatabase;
 import com.github.onedirection.database.store.EventStorer;
 import com.github.onedirection.geocoding.NamedCoordinates;
