+++ app/src/main/AndroidManifest.xml
@@ -24,7 +24,7 @@
             </intent-filter>
 
         </activity> <!-- Why on god's green earth are those lines needed for EventCreatorTest to pass ? -->
-        <activity android:name=".EventCreator" />
+        <activity android:name=".events.EventCreator" />
         <activity android:name=".EventsView" />
 
         <receiver android:name=".notifs.NotificationPublisher" />
