+++ app/src/main/java/com/github/onedirection/EventQueries.java
@@ -1,6 +1,7 @@
 package com.github.onedirection;
 
 import com.github.onedirection.database.Database;
+import com.github.onedirection.events.Event;
 
 import java.time.ZonedDateTime;
 import java.util.List;
