+++ app/src/main/java/com/github/onedirection/EventsView.java
@@ -6,10 +6,11 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.widget.TextView;
 
+import com.github.onedirection.events.Event;
+import com.github.onedirection.events.EventCreator;
 import com.github.onedirection.notifs.Notifications;
 
 import java.util.HashSet;
-import java.util.Set;
 
 public class EventsView extends AppCompatActivity {
