+++ app/src/main/java/com/github/onedirection/database/store/EventStorer.java
@@ -1,11 +1,7 @@
 package com.github.onedirection.database.store;
 
 
-import android.os.Build;
-
-import androidx.annotation.RequiresApi;
-
-import com.github.onedirection.Event;
+import com.github.onedirection.events.Event;
 import com.github.onedirection.geocoding.Coordinates;
 import com.github.onedirection.utils.Id;
