+++ app/src/main/java/com/github/onedirection/events/EventCreatorViewModel.java
@@ -1,0 +1,7 @@
+package com.github.onedirection.events;
+
+import androidx.lifecycle.ViewModel;
+
+public class EventCreatorViewModel extends ViewModel {
+    // TODO: Implement the ViewModel
+}
